Upgrading comments syntax
This has been a hard decision.
The reason why Aml used (* comments *) until now: it is what the other ML-family languages used. Retro.
Unfortunately that clashes with grammar for other language elements, such as this: `(*)` – yes, the operator reference. OCaml etc. need to write it as `( * )` just to prevent the clash.
The syntax {- comment -} does not clash with any other grammar, except maybe blocks, where a space between { and - should be used anyway. Also it goes well with the single line comment syntax. Successor ML uses (*) single line comments, which is a bit awkward, given the fact that it could refer to the * operator.